test: simplify template previewer tests
the collection media dir is necessary for running `prepareCardTextForDisplay` when a card is shown, hence why a different CollectionStorageMode was necessary

since `prepareCardTextForDisplay` isn't necessary for any of the current tests and a different CollectionStorageMode requires a heavier test configuration with Robolectric, its usage is being modified.
