+++ AnkiDroid/src/main/java/com/ichi2/anki/previewer/CardViewerViewModel.kt
@@ -18,6 +18,7 @@ package com.ichi2.anki.previewer
 import android.media.MediaPlayer
 import android.net.Uri
 import androidx.annotation.CallSuper
+import androidx.annotation.VisibleForTesting
 import androidx.core.net.toFile
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -116,7 +117,8 @@ abstract class CardViewerViewModel :
     /** From the [desktop code](https://github.com/ankitects/anki/blob/1ff55475b93ac43748d513794bcaabd5d7df6d9d/qt/aqt/reviewer.py#L358) */
     private suspend fun mungeQA(text: String) = typeAnsFilter(prepareCardTextForDisplay(text))
 
-    private suspend fun prepareCardTextForDisplay(text: String): String =
+    @VisibleForTesting
+    suspend fun prepareCardTextForDisplay(text: String): String =
         replaceAvRefsWithPlayButtons(
             text = withCol { media.escapeMediaFilenames(text) },
             renderOutput = currentCard.await().let { card -> withCol { card.renderOutput(this) } },
