+++ AnkiDroid/src/test/java/com/ichi2/anki/previewer/TemplatePreviewerViewModelTest.kt
@@ -17,19 +17,28 @@ package com.ichi2.anki.previewer
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.NotetypeFile
-import com.ichi2.anki.libanki.testutils.AnkiTest
 import com.ichi2.testutils.JvmTest
+import io.mockk.coEvery
+import io.mockk.spyk
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TemporaryFolder
 import org.junit.runner.RunWith
+import kotlin.test.assertNotEquals
 
 @RunWith(AndroidJUnit4::class)
 class TemplatePreviewerViewModelTest : JvmTest() {
     @get:Rule
     val tempDirectory = TemporaryFolder()
+
+    private fun createViewModel(arguments: TemplatePreviewerArguments): TemplatePreviewerViewModel =
+        spyk(TemplatePreviewerViewModel(arguments)).apply {
+            // the default implementation requires the Collection media directory,
+            // which needs a Robolectric setup with CollectionStorageMode.IN_MEMORY_WITH_MEDIA or ON_DISK
+            coEvery { prepareCardTextForDisplay(any()) } answers { firstArg() }
+        }
 
     @Test
     fun `getCurrentTabIndex returns the correct tab if the first cloze isn't 1 and ord isn't 0`() =
@@ -57,8 +66,19 @@ class TemplatePreviewerViewModelTest : JvmTest() {
         ) {
             onPageFinished(false)
             assertThat(this.cardsWithEmptyFronts!!.await()[0], equalTo(false))
-            assertThat(this.cardsWithEmptyFronts!!.await()[1], equalTo(true))
+            assertThat(this.cardsWithEmptyFronts.await()[1], equalTo(true))
         }
+
+    @Test
+    fun `card ords are changed`() {
+        runClozeTest(tempDirectory = tempDirectory, fields = mutableListOf("{{c1::one}} {{c2::bar}}")) {
+            onPageFinished(false)
+            val ord1 = currentCard.await().ord
+            onTabSelected(1)
+            val ord2 = currentCard.await().ord
+            assertNotEquals(ord1, ord2)
+        }
+    }
 
     private fun runClozeTest(
         ord: Int = 0,
@@ -79,25 +99,25 @@ class TemplatePreviewerViewModelTest : JvmTest() {
                 tags = mutableListOf(),
                 ord = ord,
             )
-        val viewModel = TemplatePreviewerViewModel(arguments)
+        val viewModel = createViewModel(arguments)
         block(viewModel)
     }
-}
 
-fun AnkiTest.runClozeTest(
-    ord: Int = 0,
-    tempDirectory: TemporaryFolder,
-    fields: MutableList<String>? = null,
-    block: suspend TemplatePreviewerViewModel.() -> Unit,
-) = runTest {
-    val notetype = col.notetypes.byName("Cloze")!!
-    val arguments =
-        TemplatePreviewerArguments(
-            notetypeFile = NotetypeFile(tempDirectory.root, notetype),
-            fields = fields ?: mutableListOf("{{c1::foo}} {{c2::bar}}", "anki"),
-            tags = mutableListOf(),
-            ord = ord,
-        )
-    val viewModel = TemplatePreviewerViewModel(arguments)
-    block(viewModel)
+    private fun runClozeTest(
+        ord: Int = 0,
+        tempDirectory: TemporaryFolder,
+        fields: MutableList<String>? = null,
+        block: suspend TemplatePreviewerViewModel.() -> Unit,
+    ) = runTest {
+        val notetype = col.notetypes.byName("Cloze")!!
+        val arguments =
+            TemplatePreviewerArguments(
+                notetypeFile = NotetypeFile(tempDirectory.root, notetype),
+                fields = fields ?: mutableListOf("{{c1::foo}} {{c2::bar}}", "anki"),
+                tags = mutableListOf(),
+                ord = ord,
+            )
+        val viewModel = createViewModel(arguments)
+        block(viewModel)
+    }
 }
